requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
filelock==3.13.3
Flask==3.0.2
fsspec==2024.3.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
pillow==10.2.0
sympy==1.12
torch==2.2.2
torchvision==0.17.2
typing_extensions==4.10.0
Werkzeug==3.0.1
